Add some basic integration tests and fix a bunch of bugs
Summary of changes:
- Make CommitRequest a pure data type, and hide the Deferred existence from the
User. Move the allocation to the CommitQueue.
- Move the Consumer[F] algebra to be a regular parameter rather than implicit.
It doesn't make sense as a typeclass anyway.
- Use unNoneTerminate on PartitionHandle instead of interruption. We need
graceful termination of the stream here, not abrupt interruption.
- Refactor how rebalances are handled: since the rebalance listener will be
executed on the thread running the poll, we now push rebalances into a Ref
holding a List[Rebalance], and check it after running the poll, and reset it.
- Use a shutdown queue and unNoneTerminate to stop the polling loop instead of
interruption. Using interruption caused concurrent access to the Kafka consumer.
- Use commitSync instead of commitAsync in the algebra; commitAsync will only
complete after running a poll, and this causes a deadlock on our runloop.

Resolves #4.